+++ src/Stack/Constants.hs
@@ -58,7 +58,7 @@ module Stack.Constants
     ,relDirHoogle
     ,relFileDatabaseHoo
     ,relDirPkgdb
-    ,relDirFlagCache
+    ,relFileStorage
     ,relDirLoadedSnapshotCache
     ,bindirSuffix
     ,docDirSuffix
@@ -81,7 +81,6 @@ module Stack.Constants
     ,relDirLocal
     ,relDirUsr
     ,relDirInclude
-    ,relFileDockerDb
     ,relFileIndexHtml
     ,relDirAll
     ,relFilePackageCache
@@ -94,7 +93,6 @@ module Stack.Constants
     ,relDirPackageConfInplace
     ,relDirExtraTixFiles
     ,relDirInstalledPackages
-    ,relDirPrecompiled
     ,backupUrlRelPath
     ,relDirDotLocal
     ,relDirDotSsh
@@ -399,8 +397,8 @@ relFileDatabaseHoo = $(mkRelFile "database.hoo")
 relDirPkgdb :: Path Rel Dir
 relDirPkgdb = $(mkRelDir "pkgdb")
 
-relDirFlagCache :: Path Rel Dir
-relDirFlagCache = $(mkRelDir "flag-cache")
+relFileStorage :: Path Rel File
+relFileStorage = $(mkRelFile "stack.sqlite3")
 
 relDirLoadedSnapshotCache :: Path Rel Dir
 relDirLoadedSnapshotCache = $(mkRelDir "loaded-snapshot-cached")
@@ -472,9 +470,6 @@ relDirUsr = $(mkRelDir "usr")
 relDirInclude :: Path Rel Dir
 relDirInclude = $(mkRelDir "include")
 
-relFileDockerDb :: Path Rel File
-relFileDockerDb = $(mkRelFile "docker.db")
-
 relFileIndexHtml :: Path Rel File
 relFileIndexHtml = $(mkRelFile "index.html")
 
@@ -510,9 +505,6 @@ relDirExtraTixFiles = $(mkRelDir "extra-tix-files")
 
 relDirInstalledPackages :: Path Rel Dir
 relDirInstalledPackages = $(mkRelDir "installed-packages")
-
-relDirPrecompiled :: Path Rel Dir
-relDirPrecompiled = $(mkRelDir "precompiled")
 
 backupUrlRelPath :: Path Rel File
 backupUrlRelPath = $(mkRelFile "downloaded.template.file.hsfiles")