+++ test/integration/tests/2465-init-no-packages/Main.hs
@@ -1,0 +1,11 @@
+import StackTest
+import System.Directory
+import Control.Monad (unless)
+
+main :: IO ()
+main = do
+  removeFileIgnore "stack.yaml"
+  stack ["init"]
+  exists <- doesFileExist "stack.yaml"
+  unless exists $ error "stack.yaml not created!"
+  stack ["build"]