+++ test/integration/tests/proper-rebuilds/Main.hs
@@ -11,6 +11,7 @@ main = do
         expectNoRecompilation stderr =
           when ("files-1.0.0: build" `isInfixOf` stderr) $
           error "package recompilation was not expected"
+    copyFile "src/Lib.hs.v1" "src/Lib.hs"
     stackCheckStderr ["build"] expectRecompilation
     stackCheckStderr ["build" , "--profile"] expectRecompilation
     stackCheckStderr ["build" , "--profile"] expectNoRecompilation